Unset PERM_RAW when written to
The `has_raw` variable will always be true on all writes to allocated memory, even if it has already been written to. This means that you iterate through all permissions of all writes for no reason (the second time), and this should probably only be done on the first write (at least that's what I thought the purpose of `READ_RAW` was).
For example:
If the program does 1 (kind of small) allocation of 5KB and writes twice to the whole same allocation, you are iterating 5K times for no reason per program execution. If you have 1 million executions per sec you are iterating 5 billion times for no reason since those bytes already have their read flag on them.

Also, I rechecked the vod and you misread the speedup not by checking permissions. It was about ~110x faster not just 1.2x. It went from 4.2 million fcps to 480 million fpcs.
This change will of course not give 110x speedup but it is a step in that direction without losing the permission system (so basically free).